requirements.txt
asgiref==3.3.4
bootstrap4==0.1.0
cmake==3.18.4.post1
Django==3.2
django-crispy-forms==1.11.2
django-phonenumber-field==5.1.0
phonenumbers==8.12.21
Pillow==8.2.0
pytz==2021.1
sqlparse==0.4.1
typing-extensions==3.7.4.3
